find bugs performance

+++ src/main/java/com/kancolle/server/service/battle/shelling/impl/EnmeyShipShellingSystem.java
@@ -77,7 +77,7 @@ public class EnmeyShipShellingSystem extends BaseShipShellingSystem<EnemyShip, M
     public void generateAttackList(EnemyShip attackShip, BattleContext context) {
         HougekiResult hougekiResult = context.getNowHougekiResult();
         ImmutableBiMap<Integer, IShip> shipsMap = context.getShipMap();
-        hougekiResult.getApi_at_list().add(shipsMap.inverse().get(attackShip).intValue());
+        hougekiResult.getApi_at_list().add(shipsMap.inverse().get(attackShip));
     }
 
     @Override
